add datanode list components
