+++ graylog2-web-interface/src/components/datanode/DataNodeConfiguration/CertificateRenewal.tsx
@@ -28,10 +28,11 @@ import { Spinner } from 'components/common';
 import { Alert, ListGroup, ListGroupItem, Button } from 'components/bootstrap';
 import { defaultCompare } from 'logic/DefaultCompare';
 import useSendTelemetry from 'logic/telemetry/useSendTelemetry';
-import DataNodeBadge from 'components/datanode/DataNodeBadge';
 import { Badge } from 'preflight/components/common';
 import useLocation from 'routing/useLocation';
 import { TELEMETRY_EVENT_TYPE } from 'logic/telemetry/Constants';
+
+import DataNodeBadge from '../DataNodeList/DataNodeBadge';
 
 const StyledList = styled(ListGroup)`
   max-width: fit-content;
